requirements.txt
# --- Core AI & ML ---
google-genai
xgboost
scikit-learn==1.8.0
joblib==1.5.3
numpy==2.4.4
pandas==3.0.2

# --- Data Formats ---
pyarrow==23.0.1
fastparquet==2026.3.0

# --- Utilities ---
requests==2.33.1
python-dateutil==2.9.0.post0

# --- Jupyter (Opzionale, se usi Notebook) ---
ipykernel
jupyterlab